add gateway and identity services - update auction and search services

+++ src/AuctionService/Controllers/AuctionController.cs
@@ -1,9 +1,11 @@
-
-
 using AuctionService.Data;
 using AuctionService.DTOs;
 using AuctionService.Entities;
 using AutoMapper;
+using AutoMapper.QueryableExtensions;
+using Contracts;
+using MassTransit;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 
@@ -15,21 +17,25 @@ public class AuctionController : ControllerBase
 {
     private readonly AuctionDbContext _context;
     private readonly IMapper _mapper;
-    public AuctionController(AuctionDbContext context, IMapper mapper)
+    private readonly IPublishEndpoint _publishEndpoint;
+    public AuctionController(AuctionDbContext context, IMapper mapper, IPublishEndpoint publishEndpoint)
     {
         _context = context;
         _mapper = mapper;
+        _publishEndpoint = publishEndpoint;
     }
 
     [HttpGet]
-    public async Task<ActionResult<List<AuctionDto>>> GetAllAuctions()
+    public async Task<ActionResult<List<AuctionDto>>> GetAllAuctions(string date)
     {
-        var auctions = await _context.Auctions
-        .Include(x => x.Item)
-        .OrderBy(x => x.Item.Make)
-        .ToListAsync();
+        var query = _context.Auctions.OrderBy(x => x.Item.Make).AsQueryable();
 
-        return _mapper.Map<List<AuctionDto>>(auctions);
+        if (!string.IsNullOrEmpty(date))
+        {
+            query = query.Where(x => x.UpdatedAt.CompareTo(DateTime.Parse(date).ToUniversalTime()) > 0);
+        }
+
+        return await query.ProjectTo<AuctionDto>(_mapper.ConfigurationProvider).ToListAsync();
     }
 
     [HttpGet("{id}")]
@@ -39,17 +45,20 @@ public class AuctionController : ControllerBase
         .Include(s => s.Item)
         .FirstOrDefaultAsync(s => s.Id == id);
 
+        if (auction == null) return NotFound();
+
         return _mapper.Map<AuctionDto>(auction);
     }
 
+    [Authorize]
     [HttpPost]
     public async Task<ActionResult<AuctionDto>> CreateAuction(CreateAuctionDto auctionDto)
     {
         var auction = _mapper.Map<Auction>(auctionDto);
-
-        auction.Seller = "test";
-
+        auction.Seller = User.Identity.Name;
         _context.Auctions.Add(auction);
+
+        await _publishEndpoint.Publish(_mapper.Map<AuctionCreated>(_mapper.Map<AuctionDto>(auction)));
 
         var result = await _context.SaveChangesAsync() > 0;
 
@@ -58,6 +67,7 @@ public class AuctionController : ControllerBase
         return CreatedAtAction(nameof(GetAuctionById), new { auction.Id }, _mapper.Map<AuctionDto>(auction));
     }
 
+    [Authorize]
     [HttpPut("{id}")]
     public async Task<ActionResult> UpdateAuction(Guid id, UpdateAuctionDtos updateAuctionDtos)
     {
@@ -66,11 +76,15 @@ public class AuctionController : ControllerBase
 
         if (auction == null) return NotFound();
 
+        if (auction.Seller != User.Identity.Name) return Forbid();
+
         auction.Item.Make = updateAuctionDtos.Make ?? auction.Item.Make;
         auction.Item.Model = updateAuctionDtos.Model ?? auction.Item.Model;
         auction.Item.Color = updateAuctionDtos.Color ?? auction.Item.Color;
         auction.Item.Mileage = updateAuctionDtos.Mileage ?? auction.Item.Mileage;
         auction.Item.Year = updateAuctionDtos.Year ?? auction.Item.Year;
+
+        await _publishEndpoint.Publish(_mapper.Map<AuctionUpdated>(auction));
 
         var result = await _context.SaveChangesAsync() > 0;
 
@@ -79,6 +93,7 @@ public class AuctionController : ControllerBase
         return BadRequest("Problem when saving changes");
     }
 
+    [Authorize]
     [HttpDelete("{id}")]
     public async Task<ActionResult> DeleteAuction(Guid id)
     {
@@ -86,7 +101,11 @@ public class AuctionController : ControllerBase
 
         if (auction == null) return NotFound();
 
+        if (auction.Seller != User.Identity.Name) return Forbid();
+
         _context.Auctions.Remove(auction);
+
+        await _publishEndpoint.Publish<AuctionDeleted>(new { Id = auction.Id.ToString() });
 
         var result = await _context.SaveChangesAsync() > 0;
